pyproject.toml
[project]
name = "tidepilot"
version = "0.1.0"
description = "TidePilot MCP server — exposes drafts, voice profile, and engagement data to Claude Desktop"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp>=1.0.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
tidepilot-mcp = "main:run"
